made a parent lineup fragment
